moved circle and questions and line into one folder
