+++ client/src/components/Stat.jsx
@@ -1,4 +1,4 @@
-import Questions from "./Questions";
+import Questions from "./Profile/Circle&Questions/Questions";
 import Circle from "./Circle";
 import { useState, useEffect } from "react";
 import Profile from "./Profile";
